rename RemoveConfiguratorConstantsRector to RenameClassConstantsUseToStringsRector
