+++ packages/Nette/tests/Rector/Bootstrap/RenameClassConstantsUseToStringsRector/Correct/correct.php.inc
