+++ packages/Nette/tests/Rector/Bootstrap/RenameClassConstantsUseToStringsRector/Wrong/wrong.php.inc
