+++ packages/Nette/tests/Rector/Bootstrap/RenameClassConstantsUseToStringsRector/config.yml
@@ -1,0 +1,2 @@
+services:
+    Rector\Nette\Rector\Bootstrap\RenameClassConstantsUseToStringsRector: ~
